requirements.txt
python-dotenv~=0.20.0
disnake~=2.4.0
requests~=2.27.1